kloudvin/api-python/requirements.txt
azure-functions
markitdown[docx]
azure-storage-blob
python-docx
